Avoid dup comparisons among files
Files of the same size that have the same trailer size were
being compared twice (1. When files of the same size are compared
2. When files of the same size-trailer_size are compared) if they
were not dups. Now we avoid comparing files of the same size when
comparing files with the same (size - trailer_size) .

This should increase performance by avoiding dup comparisons