pio_swapm now accepts req from usr
The caller can now provide a req to store MPI requests
corresponding to sends and recvs during rearrangement of
the data.

A util function to free the user request is also available.

Currently no invocation of pio_swapm() uses this feature.
This feature will be used later to allow a non-blocking
implementation of the pio_swapm() call.

This change also slightly reduces the memory footprint
since we only allocate memory (for storing requests)
for the number of tasks with non-zero sends/recvs.