Adding wait function for async data rearrangement
Adding a wait function that only processes asynchronous data
rearrangement.

Since we use a single queue for all asynchronous operations on
a file, it is possible to have multiple types of asynchronous
operations pending on the file. So when waiting for data
rearrangement now we only wait for asynchronous operations
corresponding to data rearrangement.