On soft close wait on iosys async ops
Now we wait on pending async iosystem ops when the iosystem is
finalized instead of waiting for pending iosystem ops when a
file is closed.

The "soft close" should only move data from compute to I/O procs
and start writing data (for PnetCDF). The waiting should now
happen when the iosystem (that included the file) is finalized.

Fixes #51