Remove file from list before I/O offload
If we are using an async write thread remove the file from the
list before offloading file operations to the async thread. This
prevents the thread from modifying the global file list (when
closing the file) and sync issues with the main thread (that might
also be operating, when opening/creating files, on the same
global list).

* The function to delete file from the global file list is now
  broken up into two functions, one that deletes (but does not free)
  the file from the global file list and a function that frees
  the file

* When using async wr thread, we first remove the file from the
  global list (delete file from list) and then offload it to the
  async thread (that eventually frees the file). In the non-async
  case the file is removed from the global list and immediately
  freed on close.