Add optional sync with io procs on close
Adding an option to decide whether to sync with I/O processes on
close.

In async mode (with an async thread), the I/O procs issue a soft
close with pending async ops and the compute procs issue a hard
close. So in this case compute procs can (and should) proceed
without syncing with I/O procs.