Moving read/write darray Fortran intefaces
Refactoring & moving read/write darray Fortran functions to a
new file.

Removing workaround required for gfortran 4.7.2 when writing
arrays since the compiler version is old (and we have newer
versions of gfortran on machines that we use). The workaround was
required due to a bug in the transfer() function with multidim
arrays.

Also removing the Fortran interface to the PIOc_write_darray_multi()
function. No applications use this interface from the Fortran
side and is currently only used internally inside the C library