Avoid zero-sized allocations in PnetCDF collective data mode
In PIOc_get_vars_tc(), a temporary count array is allocated with
ndims elements in PnetCDF collective data mode.

For scalar variables, it will be zero-sized allocations as ndims
is zero.

The C99 standard (7.20.3) states:
If the size of the space requested is zero, the behavior is
implementation-defined: either a null pointer is returned,
or the behavior is as if the size were some nonzero value,
except that the returned pointer shall not be used to access
an object.

We can avoid zero-sized allocations for scalar variables as
PnetCDF allows them to use NULL for start/count arrays.