New C function to handle error in Fortran
Adding a new function in the C library to handle error codes from
the Fortran interface. Since error codes need to be handled based
on the I/O system and file (and since this information is not
available in the Fortran interface lib) we need this new function
and its corresponding Fortran function.

Also some minor documentation changes