Avoid reserving extra space in the file header more than once
After the header section is expanded (even with sufficient free
space), a new reservation (with the same request as before) may
involve moving (shifting) data, which can be very expensive if
the data sections are huge (this is very common for output files
of some ultra-high-resolution E3SM/scream cases).

Update SCORPIO to reserve the extra header space only once.