Broadcasting the IO type to all tasks in PIOc_openfile_retry
On IO tasks, if we fail to open a file due to an incompatible type
of NetCDF, PIOc_openfile_retry might switch the IO type to plain
old basic NetCDF. Broadcast this possible update to all tasks when
the retry flag is set.

Note: This change is required only when we care about the switched
IO type on all tasks. If the updated IO type is only required on IO
tasks, this change is probably not needed.