Fix length of string attributes in puts
Fix the length of string attributes in puts by ensuring that the
extra C NULL char is not written out (while ensuring that all
strings passed to the C library are terminated with the NULL char).
This change only affects single string (not array of strings)
attributes written out using puts.