Applying some workarounds to H5DSattach_scale calls
* Place an MPI barrier to sync H5DSattach_scale calls
According to HDF5 developers, the H5DS routines are not parallel,
so all the ranks are going to be doing the same operations. At
some point, with enough iterations of the loop, HDF5 might get
out of step between the ranks.

* Disable tracking and indexing of attribute creation order
H5DSattach_scale calls (even with MPI_Barrier) might fail or hang
if attribute creation order is tracked or indexed. Before we have
a better workaround, temporarily disable tracking and indexing of
attribute creation order.