Resolving redefinition of strdup when ADIOS and HDF5 are both enabled
strdup is a function specified in the POSIX.1-2008 standard, which
defines APIs for UNIX-like operating systems. It is not part of the
ANSI C or ISO C99 standards.

Currently, pioc_support.c has its own version of strdup defined
twice, one for ADIOS type and one for HDF5 type. This results in
a redefinition error when both ADIOS and HDF5 are enabled.

To solve this problem, the custom version of strdup is renamed as
spio_strdup for clarity, and it is defined only once. This renamed
function can be used throughout the SCORPIO codebase.