Validating input file name length in openfile function
In the openfile function, include a check to ensure that the length
of the input file name, including the complete path, does not exceed
the limit defined by PIO_MAX_NAME.

This is important because exceeding this limit could cause failures
in low-level NetCDF or PnetCDF library calls, resulting in errors
such as NC_EURL (Malformed URL).

Extend the file name length validation, which was previously applied
only during createfile, to also cover openfile.

Additionally, in line with the openfile function, enhance the error
message generated during createfile to include the file name info
if the input parameters check fails.