Defining and writing attributes only on rank 0 for ADIOS IO type
For ADIOS IO type, there is no need to define the same attributes
on all ranks, which might greatly increase the attribute metadata
size and cause ADIOS lib to throw an exception when trying to
gathering more than 2^31 elements.

We can define attributes only on rank 0 to avoid repetition which
is going to be thrown away after gathering them.