Comparing iomaster with MPI_ROOT for attributes put by ADIOS type
In the past, file->adios_iomaster was employed for comparison with
MPI_ROOT, but this flag has now been entirely removed.

It is incorrect to compare file->all_rank with MPI_ROOT: only ios->
iomaster or ios->compmaster could be set to MPI_ROOT.

Opt for ios->iomaster for the comparison instead. This approach also
aligns with the handling of other IO types in PIOc_put_att_tc().