[HACK] Avoid ADIOS type when writing mpassi restart file
For restart runs of the E3SM ne4 F case, there are two problems when
attempting to read the mpassi restart file with ADIOS type:

Problem 1: During a restart run, the E3SM code initially opens the
mpassi restart file with PnetCDF type and then reopens it with ADIOS
type. If the file format is ADIOS, the first read fails. Conversely,
if the file format is PnetCDF, both reads succeed (SCORPIO switches
ADIOS type to PnetCDF type for the second read).

Problem 2: The file is written and read with different decomposition
maps, which is currently unsupported for the ADIOS type in SCORPIO.