Allocating startlist/countlist on heap in write_darray_multi_par()
When utilizing the SUBSET rearranger, the startlist and countlist
arrays may attain sizes that can lead to problems when the stack
size limits are constrained.

To address this issue, transition these arrays from stack-based to
heap-based allocation to accommodate the larger space requirements,
mitigating the risk of potential segmentation faults.