Switching to SUBSET rearranger upon local map length limit exceeded
Impose a maximum limit on the local map length for BOX rearranger.
When this limit is reached, switch to SUBSET rearranger instead.

This transition is most effective for ADIOS type, as it does not
involve data rearrangement during write. However, for PnetCDF type,
utilizing SUBSET rearranger may significantly increase write time,
potentially negating the benefits of reduced initialization time
associated with this approach.