Adding PIO_REARR_ANY rearranger
Instead of switching to SUBSET rearranger (from BOX) when a limit
(on the local I/O decomposition length) is exceeded, adding a new
rearranger that would allow this switch.

A new rearranger PIO_REARR_ANY is now available. The rearranger
chooses the optimal rearranger for a given I/O decomposition.
The range for switching to SUBSET rearranger is provided via
configure and is used for deciding whether BOX or SUBSET
rearranger is used for a specific decomposition.

Also adding a basic test for the new rearranger