Fix PIO_USE_INDEP_MODE cmake option
Fixing the PIO_USE_INDEP_MODE cmake option to allow disabling
the option on most of the machines.

Now we explicitly enable the option on all machines (i.e., use
indep mode on all machines). On Chrysalis and ANL compute nodes
we explicitly indicate that this option is overridden to use the
collective mode.

Also moving the query for the build host to the beginning of the
script (instead of querying it multiple times in the same script)