Move min/max macros to sep hdr file
Moving C min/max macros to a separate header file.

The C min/max macros cause issues with some C++ headers (conflict
with defns in STL headers). So moving these macros to a separate
header file to ease the porting of source files to C++

Also note that the header is now a C header file and the
"if defined(__cplusplus)" protections are removed for now