Figuring out the expected location of the write decomposition map
When read and write operations utilize identical decomposition
maps (as required by ADIOS type), we can determine the anticipated
start and end indices of the write decomposition map within the
target block.

These anticipated indices provide a hint for the match_decomp_part
helper function to prioritize and verify.

Additionally, they serve as default indices to read a portion of
the distributed array data buffers in case the write decomposition
map is not available.