Optionally save decomposition maps in ADIOS BP files
Storing decomposition maps in ADIOS BP files can significantly
increase disk usage. Therefore, this should be optional. A new
flag is introduced in the file structure to indicate whether to
store these maps or not, with the default set to true.

For ADIOS writers, if the flag is false, decomposition maps are
not saved. A new metadata info variable stores the flag value.

For ADIOS readers, the flag is read from BP files. If false, it
assumes that read/write operations use identical maps, and no
verification is performed.

For the ADIOS conversion tool, the flag is also read from BP
files. If false, a warning is issued, and the converted files
will exclude data from distributed array variables.