Using GPTL to determine init status
Some tools that are built with the SCORPIO library, like the
ADIOS to NetCDF conversion tool, sometimes require access to
GPTL before SCORPIO is initialized. When SCORPIO library internal
timing is enabled (for SCORPIO GPTL stats) we used to always
initialize GPTL inside the library, and this approach is not
possible when applications (like the conversion tool) require
access to the GPTL library before initializing SCORPIO (and
also require SCORPIO GPTL stats).

So we now query the GPTL library to know the library init status
before initializing (and appropriately finalizing) the lib.