Refactor initdecomp func to support 0/1 based maps
Adding the logic to initdecomp function to support 0/1 based
decomposition maps.

Separating the user interfaces (and simplifying them) to thin
wrapper functions that eventually call the initdecomp() function
to create I/O decompositions for 0 & 1 based decomps

Also using C++ algos to copy the map (and transform it as needed
to support 0/1 based maps for the different rearrangers) and
dimension lengths.