Applying a workaround for CMake FindMPI regression on Cray systems
Utilized a temporary but necessary workaround to address a confirmed
CMake regression (3.22.0 or higher) that causes find_package(MPI) to
fail or hang when detected in subprojects on Cray systems.

Until this regression is resolved by CMake developers, an additional
find_package(MPI) call is added at the root level when running CMake
on Cray systems with versions 3.22.0 or higher.