Fix issues when reopening existing files with ADIOS type
When checking whether a file is reopened with both the ADIOS type
and write mode, file->mode has not yet been assigned. As a result,
the default value 0 is always used.

Additionally, when the file description data has not been fully
populated, a NULL pointer should be used when calling pio_err().