Adding boiler plate code for pseudo E3SM cases
Generalizing the test code in test_hdf5.c and test_adios.c so
that it can be used for all I/O types.

This commit adds the boiler plate for this executable. We now
have,
* Command line arg handling (to specify I/O types, rearrangers,
  number of I/O tasks etc)
* A bunch of util functions

We do not have the F/G/I test code yet, however we have the
boiler plate functions that will be modified to add this feature